requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.6
pytest==6.1.2
six==1.14.0
allure-pytest==2.8.19
allure-python-commons==2.8.19
webdriver-manager==3.2.2
selenium
boto3
pytest-selenium
pytest-xdist
pytest-html
